+++ vmime/platforms/posix/posixHandler.hpp
@@ -55,7 +55,7 @@ public:
 	posixHandler();
 	~posixHandler();
 
-	unsigned int getUnixTime() const;
+	unsigned long getUnixTime() const;
 
 	const vmime::datetime getCurrentLocalTime() const;
